Refactor: Rename Project Setting
