+++ src/main/kotlin/org/dongguk/ownsaemiro/backlog/ReactiveApplication.kt
@@ -1,16 +1,16 @@
-package org.dongguk.ownsaemiro.reactive
+package org.dongguk.ownsaemiro.backlog
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import java.util.TimeZone
 
 @SpringBootApplication
-class ReactiveApplication {
+class BacklogApplication {
 	init {
 		TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"))
 	}
 }
 
 fun main(args: Array<String>) {
-	runApplication<ReactiveApplication>(*args)
+	runApplication<BacklogApplication>(*args)
 }
